+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   exchange rate, along with the `Instant` it was fetched. It can be initialized
   and refreshed by calling `Synchronizer.refreshExchangeRateUsd()`.
 - `ZatoshiExt.toFiatString` is now a public function
+- `Synchronizer.getFastestServers([LightWalletEndpoint])` is a flow that measures connections to given endpoints and
+  returns the three fastest ones
 
 ## [2.1.3] - 2024-08-08